+++ packages/drivekit_trip_simulator/flutter_drivekit_trip_simulator_platform_interface/CHANGELOG.md
@@ -1,7 +1,11 @@
+# 1.5.0
+
+    - Update DriveKit version to 2.12.0 for iOS and Android
+
 # 1.4.2
 
     - Remove [caret syntax](https://dart.dev/tools/pub/dependencies#caret-syntax) to use the exact DriveKit Flutter SDK version.
-
+    
 # 1.4.1
 
     - Do not run ktlint on clean and build tasks (Android)